# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Manga Scraper API server.",
        "title": "Manga Scraper API",
        "termsOfService": "https://manga-scraper.hostinger.fourleaves.studio/terms",
        "contact": {
            "name": "API Support",
            "url": "https://manga-scraper.hostinger.fourleaves.studio/support",
            "email": "admin@fourleaves.studio"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/v1/chapters/{provider_slug}/{series_slug}": {
            "get": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Get paginated chapter list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chapters"
                ],
                "summary": "Get paginated chapter list",
                "parameters": [
                    {
                        "type": "string",
                        "example": "asura",
                        "description": "Provider slug",
                        "name": "provider_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "reincarnator",
                        "description": "Series slug",
                        "name": "series_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "10",
                        "description": "Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "100",
                        "description": "Size",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/chapters/{provider_slug}/{series_slug}/_all": {
            "get": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Get all chapter list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chapters"
                ],
                "summary": "Get all chapter list",
                "parameters": [
                    {
                        "type": "string",
                        "example": "asura",
                        "description": "Provider slug",
                        "name": "provider_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "reincarnator",
                        "description": "Series slug",
                        "name": "series_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/chapters/{provider_slug}/{series_slug}/_list": {
            "get": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Get chapter list only",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chapters"
                ],
                "summary": "Get chapter list only",
                "parameters": [
                    {
                        "type": "string",
                        "example": "asura",
                        "description": "Provider slug",
                        "name": "provider_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "reincarnator",
                        "description": "Series slug",
                        "name": "series_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "v1Response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "detail": {},
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "TokenAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}